+++ dts/binder/binder-impl.d.ts
@@ -22,6 +22,7 @@ export declare class BinderImpl<T = any> implements Binder<T>, ScopedBinder, Con
     private _definesScope;
     private _createInScope;
     constructor(_primaryIdentifier: Identifier<T>);
+    readonly identifiers: Identifier[];
     to(ctor: Newable): ScopedBinder;
     toSelf(): ScopedBinder;
     toDynamicValue(factory: (context: Context) => any): ScopedBinder;